Fix golangci-lint v2 findings
Resolve the issues surfaced after upgrading to golangci-lint v2:

- client: check resp.Body.Close() error via deferred closure (errcheck)
- resource_project_test: drop unnecessary int64 conversion (unconvert)
- provider_test: remove unused testAccOrg helper (unused)
- exclude gosec G101 in *_test.go files; test fixtures use fake DSNs
  with embedded credentials (gosec)

+++ internal/provider/resource_project_test.go
@@ -93,7 +93,7 @@ func TestProjectModelFromAPI(t *testing.T) {
 			if got.DateCreated.ValueString() != tc.out.DateCreated {
 				t.Fatalf("date_created = %q, want %q", got.DateCreated.ValueString(), tc.out.DateCreated)
 			}
-			if got.EventThrottleRate.ValueInt64() != int64(tc.out.EventThrottleRate) {
+			if got.EventThrottleRate.ValueInt64() != tc.out.EventThrottleRate {
 				t.Fatalf("event_throttle_rate = %d, want %d", got.EventThrottleRate.ValueInt64(), tc.out.EventThrottleRate)
 			}
 			if tc.out.Platform == nil {
